requirements.txt
python-telegram-bot==20.3
openai>=1.0
transformers>=4.30
torch>=2.0
python-dotenv>=1.0
requests>=2.28